requirements.txt
Django==4.2.6
django-js-asset==2.1.0
django-mptt==0.15.0
flake8==6.0.0 
flake8-isort==6.0.0
sqlparse==0.4.4
python-dotenv==1.0.0